Module opendp::accuracy

Convert between noise scales and accuracies.

Functions

accuracy_to_gaussian_scale

Convert a desired accuracy (tolerance) into a gaussian noise scale at a statistical significance level alpha.

accuracy_to_laplacian_scale

Convert a desired accuracy (tolerance) into a laplacian noise scale at a statistical significance level alpha.

gaussian_scale_to_accuracy

Convert a gaussian scale into an accuracy estimate (tolerance) at a statistical significance level alpha.

laplacian_scale_to_accuracy

Convert a laplacian scale into an accuracy estimate (tolerance) at a statistical significance level alpha.