Function opendp::accuracy::accuracy_to_laplacian_scale

pub fn accuracy_to_laplacian_scale<T: Float + Zero + One + Debug>(
    accuracy: T,
    alpha: T
) -> Fallible<T>

Convert a desired accuracy (tolerance) into a laplacian noise scale at a statistical significance level alpha.

Arguments

accuracy - Desired accuracy. A tolerance for how far values may diverge from the input to the mechanism.

alpha - Statistical significance, level-alpha, or (1. - alpha)100% confidence. Must be within (0, 1].

Generics

T - Data type of accuracy and alpha

Returns

Laplacian noise scale that meets the accuracy requirement at a given level-alpha.