Module opendp::combinators

Various combinator constructors.

Traits

AmplifiableMeasure

BasicCompositionMeasure

FixDeltaMeasure

IsSizedDomain

Functions

make_basic_composition

Construct the DP composition [measurement0, measurement1, …]. Returns a Measurement that when invoked, computes [measurement0(x), measurement1(x), ...]

make_chain_mt

Construct the functional composition (measurement1 ○ transformation0). Returns a Measurement that when invoked, computes measurement1(transformation0(x)).

make_chain_tm

Construct the functional composition (transformation1 ○ measurement0). Returns a Measurement that when invoked, computes transformation1(measurement0(x)). Used to represent non-interactive postprocessing.

make_chain_tt

Construct the functional composition (transformation1 ○ transformation0). Returns a Measurement that when invoked, computes transformation1(transformation0(x)).

make_fix_delta

Fix the delta parameter in the privacy map of a measurement with a SmoothedMaxDivergence output measure.

make_population_amplification

Construct an amplified measurement from a measurement with privacy amplification by subsampling. This measurement does not perform any sampling. It is useful when you have a dataset on-hand that is a simple random sample from a larger population.

make_zCDP_to_approxDP

Constructs a new output measurement where the output measure is casted from ZeroConcentratedDivergence<QO> to SmoothedMaxDivergence<QO>.