Function opendp::combinators::make_fix_delta

pub fn make_fix_delta<DI, DO, MI, MO>(
    measurement: &Measurement<DI, DO, MI, MO>,
    delta: MO::Atom
) -> Fallible<Measurement<DI, DO, MI, MO::FixedMeasure>> where
    DI: Domain,
    DO: Domain,
    MI: 'static + Metric,
    MO: 'static + FixDeltaMeasure, 

Fix the delta parameter in the privacy map of a measurement with a SmoothedMaxDivergence output measure.

Arguments

measurement - a measurement with a privacy curve to be fixed

delta - parameter to fix the privacy curve with

Generics

DI - Input Domain

DO - Output Domain

MI - Input Metric.

MO - Output Measure of the input argument. Must be SmoothedMaxDivergence