Module opendp::core

Core concepts of OpenDP.

This module provides the central building blocks used throughout OpenDP:

Measurement

Transformation

Domain

Metric/Measure

Function

StabilityMap/PrivacyMap

Structs

Function

A mathematical function which maps values from an input Domain to an output Domain.

Measurement

A randomized mechanism with certain privacy characteristics.

PrivacyMap

A map evaluating the privacy of a Measurement.

StabilityMap

A map evaluating the stability of a Transformation.

Transformation

A data transformation with certain stability characteristics.

Traits

DatasetMetric

An indicator trait that is only implemented for dataset distances.

Domain

A set which constrains the input or output of a Function.

Measure

A representation of the distance between two distributions.

Metric

A representation of the distance between two elements in a set.

SensitivityMetric

An indicator trait that is only implemented for statistic distances.