Trait opendp::core::Metric

pub trait Metric: Default + Clone + PartialEq + Debug {
    type Distance;
}

A representation of the distance between two elements in a set.

Proof Definition

A type Self has an implementation for Metric if an only if it can represent a metric for quantifying distances between values in a set.

Required Associated Types

type Distance

Proof Definition

Self::Distance is a type that represents distances in terms of a metric Self.