Module opendp::domains

Various implementations of Domain.

These different versions of Domain provide a general set of models used throughout OpenDP. Most of the implementations are generic, with the type parameter setting the underlying Domain::Carrier type.

A data domain is a representation of the set of values on which the function associated with a transformation or measurement can operate. Each metric (see crate::metrics) is associated with certain data domains. The Domain trait is implemented for all domains used in OpenDP.

Structs

AllDomain

Proof Definition

BoundedDomain

Proof Definition

BoxDomain

A Domain that carries an underlying Domain in a Box.

DataDomain

A Domain that unwraps a Data wrapper.

InherentNullDomain

Proof Definition

MapDomain

A Domain that contains maps of (homogeneous) values.

OptionNullDomain

A domain that represents nullity via the Option type.

PairDomain

A Domain that contains pairs of values.

PolyDomain

A polymorphic Domain. This admits any value of any type (represented as a Box).

SizedDomain

A Domain that specifies the length of the enclosed domain.

VectorDomain

A Domain that contains vectors of (homogeneous) values.

Traits

CollectionDomain