Function opendp::measurements::make_base_discrete_laplace_linear

pub fn make_base_discrete_laplace_linear<D, QO>(
    scale: QO,
    bounds: Option<(D::Atom, D::Atom)>
) -> Fallible<Measurement<D, D, D::InputMetric, MaxDivergence<QO>>> where
    D: DiscreteLaplaceDomain,
    D::Atom: Integer + SampleDiscreteLaplaceLinear<QO>,
    QO: Float + InfCast<D::Atom>, 

Make a Measurement that adds noise from the discrete_laplace(scale) distribution to the input, using a linear-time algorithm on finite data types.

This algorithm can be executed in constant time if bounds are passed. Set D to change the input data type and input metric:

| `D`                        | input type | `D::InputMetric`     |
| -------------------------- | ---------- | -------------------- |
| AllDomain<T> (default)     | T          | AbsoluteDistance<T>  |
| VectorDomain<AllDomain<T>> | Vec<T>     | L1Distance<T>        |

Citations

GRS12 Universally Utility-Maximizing Privacy Mechanisms

Arguments

scale - Noise scale parameter for the distribution. scale == sqrt(2) * standard_deviation.

bounds - Set bounds on the count to make the algorithm run in constant-time.

Generics

D - Domain of the data type to be privatized. Valid values are VectorDomain<AllDomain<T>> or AllDomain<T>

QO - Data type of the scale and output distance.