Function opendp::measurements::make_base_geometric

pub fn make_base_geometric<D, QO>(
    scale: QO,
    bounds: Option<(D::Atom, D::Atom)>
) -> Fallible<Measurement<D, D, D::InputMetric, MaxDivergence<QO>>> where
    D: DiscreteLaplaceDomain,
    D::Atom: Integer + SampleDiscreteLaplaceLinear<QO>,
    QO: Float + InfCast<D::Atom>, 

👎 Deprecated since 0.5.0:
Use make_base_discrete_laplace instead. For a constant-time algorithm, pass bounds into make_base_discrete_laplace_linear.

Deprecated. Use make_base_discrete_laplace instead (more efficient). make_base_discrete_laplace_linear has a similar interface with the optional constant-time bounds.

Arguments

scale - Noise scale parameter for the distribution. scale == sqrt(2) * standard_deviation.

bounds - Set bounds on the count to make the algorithm run in constant-time.

Arguments

D - Domain of the data type to be privatized. Valid values are VectorDomain<AllDomain<T>> or AllDomain<T>

QO - Data type of the scale and output distance