Module opendp::measures

Various implementations of Measures (and associated Distance).

A Measure is used to measure the distance between distributions. The distance is expressed in terms of an associated type.

Example

MaxDivergence<Q> has an associated distance type of Q. This means that the symmetric distance between vectors is expressed in terms of the type Q. In this context Q is usually f32 or f64.

Structs

FixedSmoothedMaxDivergence

$(\epsilon, \delta)$-approximate differential privacy.

MaxDivergence

$\epsilon$-pure differential privacy.

SMDCurve

A function mapping from $\delta$ to $\epsilon$.

SmoothedMaxDivergence

$\epsilon(\delta)$-approximate differential privacy.

ZeroConcentratedDivergence

$\rho$-zero concentrated differential privacy.