Module opendp::traits::samplers

Traits for sampling from probability distributions.

Traits

CastInternalRational

Casting between floating-point and rational values.

SampleBernoulli

Sample a single bit with arbitrary probability of success

SampleDiscreteGaussianZ2k

Sample from the discrete gaussian distribution on $\mathbb{Z} \cdot 2^k$.

SampleDiscreteLaplaceLinear

Sample from the censored two-sided geometric distribution.

SampleDiscreteLaplaceZ2k

Sample from the discrete laplace distribution on $\mathbb{Z} \cdot 2^k$.

SampleGeometric

Sample from the censored geometric distribution.

SampleStandardBernoulli

Sample from Bernoulli(p=0.5).

SampleUniform

Sample exactly from the uniform distribution.

SampleUniformInt

Sample an integer uniformly over [Self::MIN, upper].

SampleUniformIntBelow

Sample an integer uniformly over [Self::MIN, upper]

Shuffle

Shuffle a mutable reference to a collection.

Functions

fill_bytes

Fill a byte buffer with random bits.

sample_discrete_gaussian

Sample exactly from the discrete gaussian distribution with arbitrary precision.

sample_discrete_laplace

Sample exactly from the discrete laplace distribution with arbitrary precision.