Function opendp::traits::samplers::fill_bytes

pub fn fill_bytes(buffer: &mut [u8]) -> Fallible<()>

Fill a byte buffer with random bits.

Proof Definition

For any input buffer, fill the buffer with random bits, where each bit is an iid draw from Bernoulli(p=0.5). Return Err(e) if there is insufficient system entropy, otherwise return Ok(()).