Function opendp::traits::samplers::sample_discrete_gaussian

pub fn sample_discrete_gaussian(scale: Rational) -> Fallible<Integer>

Sample exactly from the discrete gaussian distribution with arbitrary precision.

Proof Definition

For any scale that is a non-negative rational number, return Ok(x) where x is a sample from the discrete_gaussian(scale) distribution, or Err(e), due to lack of system entropy.

Specifically, the probability of returning any x of type rug::Integer is

\forall x \in \mathbb{Z} \quad  
P[X = x] = \frac{e^{-\frac{x^2}{2\sigma^2}}}{\sum_{y\in\mathbb{Z}}e^{-\frac{y^2}{2\sigma^2}}} \quad 
\text{where } X \sim \mathcal{N}_\mathbb{Z}(0, \sigma^2)

where $\sigma = scale$.

Citation

CKS20 The Discrete Gaussian for Differential Privacy