Trait opendp::traits::samplers::SampleDiscreteLaplaceZ2k

pub trait SampleDiscreteLaplaceZ2k: Sized {
    fn sample_discrete_laplace_Z2k(
        shift: Self,
        scale: Self,
        k: i32
    ) -> Fallible<Self>;
}

Sample from the discrete laplace distribution on $\mathbb{Z} \cdot 2^k$.

Implemented for floating-point types f32 and f64.

k can be chosen to be very negative, to get an arbitrarily fine approximation to continuous laplacian noise.

Required Methods

fn sample_discrete_laplace_Z2k(shift: Self, scale: Self, k: i32) -> Fallible<Self>

Proof Definition

For any setting of the input arguments, return either Err(e) if there is insufficient system entropy, or Ok(sample), where sample is distributed according to a modified discrete_laplace(shift, scale).

The modifications to the discrete laplace are as follows:

the shift is rounded to the nearest multiple of $2^k$

the sample is rounded to the nearest value of type Self.

the noise granularity is in increments of $2^k$.