Trait opendp::traits::samplers::Shuffle

pub trait Shuffle {
    fn shuffle(&mut self) -> Fallible<()>;
}

Shuffle a mutable reference to a collection.

Required Methods

fn shuffle(&mut self) -> Fallible<()>

Proof Definition

For any input self of type Self, mutate self such that the elements within are ordered randomly. Returns Err(e) if there is insufficient system entropy, or Ok(()) otherwise.