Trait opendp::traits::CheckNull

pub trait CheckNull {
    fn is_null(&self) -> bool;
}

Checks if a value is null.

Since crate::domains::AllDomain only includes non-null values, this trait is necessary for its member check.

Required Methods

fn is_null(&self) -> bool

Proof Definition

For any value of type Self, returns true if is null, otherwise false.