Trait opendp::traits::FloatBits

pub trait FloatBits: Copy + Sized + ExactIntCast<Self::Bits> {
    type Bits: Copy + One + Zero + Eq + Shr<Output = Self::Bits> + Shl<Output = Self::Bits> + BitAnd<Output = Self::Bits> + BitOr<Output = Self::Bits> + Sub<Output = Self::Bits> + From<bool>;

    const EXPONENT_BITS: Self::Bits;
    const MANTISSA_BITS: Self::Bits;
    const EXPONENT_BIAS: Self::Bits;

    fn to_bits(self) -> Self::Bits;
    fn from_bits(bits: Self::Bits) -> Self;

    fn sign(self) -> bool { ... }
    fn raw_exponent(self) -> Self::Bits { ... }
    fn mantissa(self) -> Self::Bits { ... }
    fn from_raw_components(
        sign: bool,
        raw_exponent: Self::Bits,
        mantissa: Self::Bits
    ) -> Self { ... }
    fn to_raw_components(self) -> (bool, Self::Bits, Self::Bits) { ... }
}

Bitwise consts and decompositions for float types.

Required Associated Types

type Bits: Copy + One + Zero + Eq + Shr<Output = Self::Bits> + Shl<Output = Self::Bits> + BitAnd<Output = Self::Bits> + BitOr<Output = Self::Bits> + Sub<Output = Self::Bits> + From<bool>

Proof Definition

An associated type that captures the bit representation of Self.

Required Associated Constants

const EXPONENT_BITS: Self::Bits

Proof Definition

A constant equal to the number of bits in exponent.

const MANTISSA_BITS: Self::Bits

Proof Definition

A constant equal to the number of bits in mantissa.

Note

This should be equal to Self::MANTISSA_DIGITS - 1, because of the implied leading bit.

const EXPONENT_BIAS: Self::Bits

Proof Definition

A constant equal to the bias correction of the exponent.

Required Methods

fn to_bits(self) -> Self::Bits

Proof Definition

For any self of type Self, retrieves the corresponding bit representation.

fn from_bits(bits: Self::Bits) -> Self

Proof Definition

For any bits of the associated Bits type, returns the floating-point number corresponding to bits.

Provided Methods

fn sign(self) -> bool

Proof Definition

For any self of type Self, returns true if self is positive, otherwise false.

fn raw_exponent(self) -> Self::Bits

Proof Definition

For any self of type Self, returns the bits representing the exponent, without bias correction.

fn mantissa(self) -> Self::Bits

Proof Definition

For any self of type Self, returns the bits representing the mantissa.

fn from_raw_components(
sign: bool,
raw_exponent: Self::Bits,
mantissa: Self::Bits
) -> Self

Proof Definition

For any set of arguments, returns the corresponding floating-point number according to IEEE-754.

Notes

Assumes that the bits to the left of the bit range of the raw_exponent and mantissa are not set.

fn to_raw_components(self) -> (bool, Self::Bits, Self::Bits)

Proof Definition

For any self of type Self, decomposes the type into the sign, raw exponent and mantissa.