Trait opendp::traits::InfMul

pub trait InfMul: Sized + AlertingMul {
    fn inf_mul(&self, v: &Self) -> Fallible<Self>;
    fn neg_inf_mul(&self, v: &Self) -> Fallible<Self>;
}

Fallible multiplication with specified rounding.

Throws an error if the ideal output is not finite or representable.

Required Methods

fn inf_mul(&self, v: &Self) -> Fallible<Self>

Proof Definition

For any two values self and v of type Self, self.inf_mul(v) either returns Ok(out), where $out \ge self \cdot v$, or Err(e).

fn neg_inf_mul(&self, v: &Self) -> Fallible<Self>

Proof Definition

For any two values self and v of type Self, self.neg_inf_mul(v) either returns Ok(out), where $out \le self \cdot v$, or Err(e).