Trait opendp::traits::TotalOrd

pub trait TotalOrd: PartialOrd + Sized {
    fn total_cmp(&self, other: &Self) -> Fallible<Ordering>;

    fn total_max(self, other: Self) -> Fallible<Self> { ... }
    fn total_min(self, other: Self) -> Fallible<Self> { ... }
    fn total_clamp(self, min: Self, max: Self) -> Fallible<Self> { ... }
    fn total_lt(&self, other: &Self) -> Fallible<bool> { ... }
    fn total_le(&self, other: &Self) -> Fallible<bool> { ... }
    fn total_gt(&self, other: &Self) -> Fallible<bool> { ... }
    fn total_ge(&self, other: &Self) -> Fallible<bool> { ... }
}

TotalOrd is well-defined on types that are Ord on their non-null values.

The framework provides a way to ensure values are non-null at runtime. This trait should only be used when the framework can rely on these assurances. TotalOrd shares the same interface as Ord, but with a total_ prefix on methods

Required Methods

fn total_cmp(&self, other: &Self) -> Fallible<Ordering>

Proof Definition

For any two values self and other of type Self, returns Ok(out) or Err(e). The implementation returns Err(e) if either self or other are null. Otherwise out is the Ordering of self and other as defined by PartialOrd.

Provided Methods

fn total_max(self, other: Self) -> Fallible<Self>

Proof Definition

For any two values self and other of type Self, returns Ok(out) or Err(e). The implementation returns Err(e) if either self or other are null. Otherwise returns Some(out) where out is the greater of self and other as defined by PartialOrd.

fn total_min(self, other: Self) -> Fallible<Self>

Proof Definition

For any two values self and other of type Self, returns Ok(out) or Err(e). The implementation returns Err(e) if either self or other are null. Otherwise returns Some(out) where out is the lesser of self and other as defined by PartialOrd.

fn total_clamp(self, min: Self, max: Self) -> Fallible<Self>

Proof Definition

For any three values self, min and max of type Self, returns Ok(out) or Err(e). The implementation returns Err(e) if any of self, min or max are null. Otherwise returns Some(out) where out is min if $self \lt min$, max if $self \gt max$, or else self.

fn total_lt(&self, other: &Self) -> Fallible<bool>

Proof Definition

For any two values self and other of type Self, returns Ok(out) or Err(e). The implementation returns Err(e) if either self or other are null. Otherwise returns Ok(out) where out is true if $self \lt other$.

fn total_le(&self, other: &Self) -> Fallible<bool>

Proof Definition

For any two values self and other of type Self, returns Ok(out) or Err(e). The implementation returns Err(e) if either self or other are null. Otherwise returns Ok(out) where out is true if $self \le other$.

fn total_gt(&self, other: &Self) -> Fallible<bool>

Proof Definition

For any two values self and other of type Self, returns Ok(out) or Err(e). The implementation returns Err(e) if either self or other are null. Otherwise returns Ok(out) where out is true if $self \gt other$.

fn total_ge(&self, other: &Self) -> Fallible<bool>

Proof Definition

For any two values self and other of type Self, returns Ok(out) or Err(e). The implementation returns Err(e) if either self or other are null. Otherwise returns Ok(out) where out is true if $self \ge other$.