Function opendp::transformations::make_sized_bounded_sum

pub fn make_sized_bounded_sum<MI, T>(
    size: usize,
    bounds: (T, T)
) -> Fallible<Transformation<SizedDomain<VectorDomain<BoundedDomain<T>>>, AllDomain<T>, MI, AbsoluteDistance<T>>> where
    MI: Metric,
    T: MakeSizedBoundedSum<MI>, 

Make a Transformation that computes the sum of bounded data with known dataset size.

This uses a restricted-sensitivity proof that takes advantage of known dataset size for better utility. Use make_clamp to bound data and make_bounded_resize to establish dataset size.

Citations

CSVW22 Widespread Underestimation of Sensitivity…

DMNS06 Calibrating Noise to Sensitivity in Private Data Analysis

Arguments

size - Number of records in input data.

bounds - Tuple of lower and upper bounds for data in the input domain.

Generics

MI - Input Metric. One of SymmetricDistance or InsertDeleteDistance.

T - Atomic Input Type and Output Type.