Trait opendp::transformations::DropNullDomain

pub trait DropNullDomain: Domain {
    type Imputed;

    fn option(value: &Self::Carrier) -> Option<Self::Imputed>;
}

Utility trait to drop null values from a dataset, regardless of the representation of nullity.

Required Associated Types

type Imputed

This is the type of Self::Carrier after dropping null.

On any type D for which the DropNullDomain trait is implemented, the syntax D::Imputed refers to this associated type. For example, consider D to be OptionNullDomain<T>, the domain of all Option<T>. The implementation of this trait for DropNullDomain<T> designates that type Imputed = T. Thus DropNullDomain<T>::Imputed is T.

Required Methods

fn option(value: &Self::Carrier) -> Option<Self::Imputed>

Standardizes D::Carrier into an Option<D::Imputed>, where D::Imputed is never null.

Self::Imputed may have the capacity to represent null (like f64), but implementations of this function must guarantee that Self::Imputed is never null.