Trait opendp::transformations::ImputeConstantDomain

pub trait ImputeConstantDomain: Domain {
    type Imputed;

    fn impute_constant<'a>(
        default: &'a Self::Carrier,
        constant: &'a Self::Imputed
    ) -> &'a Self::Imputed;
}

Utility trait to impute with a constant, regardless of the representation of nullity.

Required Associated Types

type Imputed

This is the type of Self::Carrier after imputation.

On any type D for which the ImputeConstantDomain trait is implemented, the syntax D::Imputed refers to this associated type. For example, consider D to be OptionNullDomain<T>, the domain of all Option<T>. The implementation of this trait for OptionNullDomain<T> designates that type Imputed = T. Thus OptionNullDomain<T>::Imputed is T.

Proof Definition

Self::Imputed can represent the set of possible output values after imputation.

Required Methods

fn impute_constant<'a>(
default: &'a Self::Carrier,
constant: &'a Self::Imputed
) -> &'a Self::Imputed

A function that replaces a potentially-null carrier type with a non-null imputed type.

Proof Definition

For any setting of the input parameters, where constant is non-null, the function returns a non-null value.