Module opendp::metrics

Various implementations of Metrics (and associated Distance).

A Metric is used to measure the distance between data. Metrics are paired with a domain on which the metric can measure distance. The distance is expressed in terms of an associated type.

Example

SymmetricDistance can be paired with a domain: VectorDomain(AllDomain(T)). In this context, the SymmetricDistance is used to measure the distance between any two vectors of elements of type T. The SymmetricDistance has an associated distance type of u32. This means that the symmetric distance between vectors is expressed in terms of a u32.

Structs

AbsoluteDistance

The absolute distance between two vector-valued aggregates.

AgnosticMetric

A dummy to fill the metric position in postprocessors.

ChangeOneDistance

The smallest number of changes to make two equal-length datasets equivalent.

DiscreteDistance

Indicates if two elements are equal to each other.

HammingDistance

The number of elements that differ between two equal-length datasets.

InsertDeleteDistance

The smallest number of insertions or deletions to make two datasets equivalent.

LpDistance

The $L_p$ distance between two vector-valued aggregates.

SymmetricDistance

The smallest number of additions or removals to make two datasets equivalent.

Type Definitions

IntDistance

The type that represents the distance between datasets. It is used as the associated Metric::Distance type for e.g. SymmetricDistance, InsertDeleteDistance, etc.

L1Distance

The $L_1$ distance between two vector-valued aggregates.

L2Distance

The $L_2$ distance between two vector-valued aggregates.