Trait opendp::traits::InfCast

pub trait InfCast<TI>: Sized {
    fn inf_cast(v: TI) -> Fallible<Self>;
    fn neg_inf_cast(v: TI) -> Fallible<Self>;
}

Fallible casting where the casted value rounds towards infinity.

This preserves the invariant that the casted value is gte the original value. For example, casting a 128_u8 to i8 doesn’t saturate to i8::MAX (127), it errors.

Required Methods

fn inf_cast(v: TI) -> Fallible<Self>

Proof Definition

For any v of type TI, Self::inf_cast(value) either returns Ok(out) or Err(e), where $out \ge v$.

fn neg_inf_cast(v: TI) -> Fallible<Self>

Proof Definition

For any v of type TI, Self::inf_cast(value) either returns Ok(out) or Err(e), where $out \le v$.